+++ exercise-reactjs/src/app/core/interfaces/ArticleInterface.ts
@@ -1,0 +1,14 @@
+export interface IArticle {
+  id: number;
+  title: string;
+  desc: string;
+  author: string;
+  createdAt: string;
+  minsRead: string;
+  category: string;
+  image: string;
+}
+
+export type IArticleDetail = IArticle & {
+  content: string;
+}
