+++ exercise-reactjs/src/app/pages/article/containers/Article.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { IArticle } from './ArticleList';
 import { BY, MINSREAD, EST } from '../../../core/constants/text';
 import { functions } from '../../../core/functions';
 import { Link } from 'react-router-dom';
+import { IArticle } from '../../../core/interfaces/ArticleInterface';
 
 interface PropsInterface {
   article: IArticle;
