+++ exercise-reactjs/src/app/store/articles/reducer.ts
@@ -1,4 +1,4 @@
-import { IAction } from '../interface';
+import { IAction } from '../../core/interfaces/ReduxInterface';
 import * as types from './types';
 
 const initialState = {
