+++ exercise-reactjs/src/app/store/rootReducer.ts
@@ -1,5 +1,5 @@
-import { combineReducers } from "redux";
-import ArticlesReducer from "./articles/reducer";
+import { combineReducers } from 'redux';
+import ArticlesReducer from './articles/reducer';
 
 const rootReducer = combineReducers({
   articles: ArticlesReducer,
