+++ exercise-reactjs/src/app/store/store.ts
@@ -1,6 +1,6 @@
-import { applyMiddleware, createStore } from "redux";
-import thunk from "redux-thunk";
-import rootReducer from "./rootReducer";
+import { applyMiddleware, createStore } from 'redux';
+import thunk from 'redux-thunk';
+import rootReducer from './rootReducer';
 
 const store = createStore(
   rootReducer,
